some improvements

+++ app/models/farm_template.py
@@ -8,9 +8,9 @@ class FarmTemplate(Base):
     farm_id = Column(Integer, primary_key=True)
     name = Column(String(100), nullable=False)
     image_url = Column(String(500), nullable=False)
-    base_price = Column(Float, nullable=False)
+    base_price = Column(Integer, nullable=False)
     price_multiplier = Column(Float, nullable=False)
-    base_income = Column(Float, nullable=False)
+    base_income = Column(Integer, nullable=False)
     income_multiplier = Column(Float, nullable=False)
     is_visible = Column(Boolean, default=False)
